Added missing cast, removed deprecated drawers

+++ src/main/java/songscribe/Proba.java
@@ -74,7 +74,7 @@ public class Proba {
             setPreferredSize(new Dimension(600, 600));
 
             try {
-                fughetta = Font.createFont(Font.TRUETYPE_FONT, new File("fonts/fughetta.ttf")).deriveFont(s);
+                fughetta = Font.createFont(Font.TRUETYPE_FONT, new File("fonts/fughetta.ttf")).deriveFont((float) s);
             }
             catch (Exception e) {
                 e.printStackTrace();
